refactor structure code
