+++ fe/src/App.tsx
@@ -15,14 +15,14 @@ import AdLicenseDetail from "pages/admin/AdLicense/components/AdLicenseDetail";
 import DistrictManagement from "pages/admin/DistrictManagement";
 import WardManagement from "pages/admin/WardManagement";
 import { routes } from "routes/routes";
-import LocationManagement from "pages/admin/ward/pages/LocationsManagement";
-import { LocationEdit } from "pages/admin/ward/pages/LocationEdit";
-import AdvertisesOfLocationManagement from "pages/admin/ward/pages/AdvertisesOfLocationManagement";
-import { AdvertiseDetail } from "pages/admin/ward/pages/AdvertiseDetail";
-import { AdvertiseEdit } from "pages/admin/ward/pages/AdvertiseEdit";
-import ReportsManagement from "pages/admin/ward/pages/ReportsManagement";
-import { ReportHandle } from "pages/admin/ward/pages/ReportHandle";
-import { ReportDetail } from "pages/admin/ward/pages/ReportDetail";
+import LocationManagement from "pages/admin/LocationsManagement";
+import { LocationEdit } from "pages/admin/LocationEdit";
+import AdvertisesOfLocationManagement from "pages/admin/AdvertisesOfLocationManagement";
+import { AdvertiseDetail } from "pages/admin/AdvertiseDetail";
+import { AdvertiseEdit } from "pages/admin/AdvertiseEdit";
+import ReportsManagement from "pages/admin/ReportsManagement";
+import { ReportHandle } from "pages/admin/ReportHandle";
+import { ReportDetail } from "pages/admin/ReportDetail";
 import CreateAccount from "pages/admin/CreateAccount";
 import ContractDetail from "pages/admin/ContractDetail";
 import ReportFormManagement from "pages/admin/ReportFormManagement";
