+++ fe/src/advertises-of-location.json
