+++ fe/src/components/common/Search/styles.module.scss
