+++ fe/src/components/common/TableTemplate/styles.module.scss
