+++ fe/src/locations.json
