+++ fe/src/pages/admin/AdvertiseDetail/components/InfoContract/index.tsx
