+++ fe/src/pages/admin/AdvertiseDetail/components/InfoContract/styles.module.scss
