+++ fe/src/pages/admin/LocationsManagement/styles.module.scss
