+++ fe/src/pages/admin/ReportDetail/styles.module.scss
