+++ fe/src/pages/admin/ReportsManagement/index.tsx
@@ -1,21 +1,22 @@
 import React, { useState, useEffect } from "react";
-import { Pagination, Box } from "@mui/material";
+import { Box, Pagination } from "@mui/material";
+
 import { useNavigate, useLocation, useResolvedPath, createSearchParams } from "react-router-dom";
 import queryString from "query-string";
 
-import classes from "./styles.module.scss";
-
 import SideBarWard from "components/admin/SidebarWard";
-import TableTemplate from "../../components/common/TableTemplate";
-import SearchAppBar from "../../components/common/Search";
-import { Header } from "../../components/common/Header";
-import LocationService from "services/location";
+import TableTemplate from "components/common/TableTemplate";
+import SearchAppBar from "components/common/Search";
+import { Header } from "components/common/Header";
+import classes from "./styles.module.scss";
+import ReportService from "services/report";
 import { routes } from "routes/routes";
 
-const LocationManagement = () => {
+const ReportsManagement = () => {
   const navigate = useNavigate();
   const itemsPerPage = 5;
-  const [locationList, setLocationList] = useState([]);
+  const [reportList, setReportList] = useState([]);
+
   const [searchValue, setSearchValue] = useState("");
 
   const locationHook = useLocation();
@@ -38,14 +39,10 @@ const LocationManagement = () => {
   };
 
   useEffect(() => {
-    const getAllLocations = async () => {
-      LocationService.getLocations({
-        search: searchValue,
-        pageSize: itemsPerPage,
-        current: Number(currentPage)
-      })
+    const getAllReports = async () => {
+      ReportService.getReports({ search: searchValue, pageSize: itemsPerPage, current: Number(currentPage) })
         .then((res) => {
-          setLocationList(res.content);
+          setReportList(res.content);
           setTotalPage(res.totalPages);
           setTotalElements(res.totalElements);
 
@@ -60,47 +57,41 @@ const LocationManagement = () => {
           console.log(e);
         });
     };
-    getAllLocations();
+    getAllReports();
   }, [currentPage, searchValue]);
 
   useEffect(() => {
     setCurrentPage(1);
   }, [searchValue]);
 
-  const data = locationList.map((location: any, index: number) => {
+  const data = reportList.map((report: any, index: number) => {
     return {
       stt: index + 1,
-      id: location.id,
-      address: location.address,
-      adsForm: location.adsForm.name,
-      objectStatus: {
-        name: location.planning ? "Đã quy hoạch" : "Chưa quy hoạch",
-        value: location.planning
-      }
+      objectStatus: { value: report.status, name: report.status ? "Đã xử lí" : "Chưa xử lí" },
+      ...report
     };
   });
 
-  const customHeading = ["STT", "Mã", "Địa chỉ", "Hình thức quảng cáo", "Tình trạng quy hoạch"];
-  const customColumns = ["stt", "id", "address", "adsForm", "objectStatus"];
-
-  // const paginatedData = data.slice((page - 1) * itemsPerPage, page * itemsPerPage);
-
-  const handleViewAds = (idLocation: number) => {
-    navigate(`${routes.admin.advertises.ofLocation.replace(":id", `${idLocation}`)}`);
-  };
-
-  const handleEditLocation = (idLocation: number) => {
-    navigate(`${routes.admin.locations.edit.replace(":id", `${idLocation}`)}`);
-  };
+  const customHeading = ["STT", "Mã", "Email", "Tên", "Điện thoại", "Tình trạng xử lý"];
+  const customColumns = ["stt", "id", "email", "fullName", "phone", "objectStatus"];
 
   const handleSearch = (query: string) => {
     setSearchValue(query);
+  };
+  // const paginatedData = data.slice((page - 1) * itemsPerPage, page * itemsPerPage);
+
+  const handleReport = (idReport: number) => {
+    navigate(`${routes.admin.reports.edit.replace(":id", `${idReport}`)}`);
+  };
+
+  const handleViewDetails = (idReport: number) => {
+    navigate(`${routes.admin.reports.details.replace(":id", `${idReport}`)}`);
   };
 
   return (
     <Box>
       <Header />
-      <div className={classes["location-management-container"]}>
+      <div className={classes["reports-management-container"]}>
         <SideBarWard></SideBarWard>
         <Box className={classes["container-body"]}>
           <Box className={classes["search-container"]}>
@@ -113,8 +104,8 @@ const LocationManagement = () => {
                 customHeading={customHeading}
                 customColumns={customColumns}
                 isActionColumn={true}
-                onViewAdsClick={handleViewAds}
-                onEditClick={handleEditLocation}
+                onEditClick={handleReport}
+                onViewDetailsClick={handleViewDetails}
               />
 
               <Box className={classes["pagination-custom"]}>
@@ -132,4 +123,4 @@ const LocationManagement = () => {
   );
 };
 
-export default LocationManagement;
+export default ReportsManagement;
