--- fe/src/pages/admin/ward/models/Location.model.ts
@@ -1,10 +1,0 @@
-export interface Location {
-  stt: number;
-  id: number;
-  address: string;
-  adsForm: string;
-  locationType: string;
-  planning: number;
-  latitude: number;
-  longtitude: number;
-}
