+++ fe/src/report-detail.json
